actions: modify mechaism for populating action groups in CAGandUIManager
Use a meta class to find all super classes that define a populate_action_groups
method and then invoke them all before adding the UI description to the UI
manager. This "additive" approach to populating the groups mean it will be
possible to create mixins that add actions without having to be descendants
of CAGandUIManager.

Signed-off-by: Peter Williams <pwil3058@gmail.com>